migrated code to use test.MesosTestCase
