+++ src/main/test/net/elodina/mesos/dse/ClusterTest.scala
@@ -5,7 +5,7 @@ import org.junit.Assert._
 import net.elodina.mesos.dse.Util.BindAddress
 import net.elodina.mesos.util.Range
 
-class ClusterTest extends MesosTestCase {
+class ClusterTest extends DseMesosTestCase {
   @Before
   override def before {
     super.before
